Fix spurious trigger 2 press on startup
We are unfortunately using pin 13 of Arduino Nano for controlling trigger 2.
That pin is hardwired to the builtin led which causes it to lit (slightly) when
trigger 2 is not pressed and thus pulled high by the MSX side.
It also causes a spurious press on startup.

Instead of configuring the pin as input (high impedance) when trigger 2
is not pressed, configure the pin as output and set it to high. When
trigger 2 is pressed, the pin is configured as output and set to low as
before.

This makes de led light except when trigger 2 is pressed and avoids the
spurious trigger on startup.

This behaviour was detected when using the msx-joydolphin-v1 adapter
on a MSXVR with Hell's Road: when starting the game, or when plugging
the adapter, a bomb was always detonated (which is done by pressing
trigger 2).

Also, the controller initialization delay is removed (testing shows no
change in behavior without the delay) and the TX led is no longer lit
when a working controller is detected and buttons or pads are pressed.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>