firmware: revert to ignore pin8
Due to the strict timings required to read the gamecube controller
interrupts are disabled during those periods which lead to
losing interrupts related to pin8 changes.

Although the previous code passed the HIDtest test, observing
the signals with a osciloscope reveals that some times pin8 is
not honored correctly because of the lost pin8 change interrupts.

As ignoring pin8 does not cause a change in behaviour when pin8 is
low (as required to get the status of joysticks) vs a standard MSX
joystick, we keep it this way.

Signed-off-by: Albert Herranz <albert_herranz@yahoo.es>